8.3.0
added new features, fixed a many bugs, cleaned/shrinked a lot of code

+++ commands/Admin/broadcast.js
@@ -8,13 +8,6 @@ module.exports = {
     usage: "broadcast\n**e.g.**\n\`broadcast\`\n> Follow the instructions\n> Bot will ask you a title and a description/message\n> Title can be max 256 char long, Description can be max 2048 char long",
     permission: "ADMINISTRATOR",
     run: async (client, msg, arg) => {
-        if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
-
-        let author = msg.author
-
-        var filter = m => m.author.id === msg.author.id;
-        var title;
-        var description;
 
         const nobotpermEmbed = new Discord.MessageEmbed()
             .setColor(`RED`)
@@ -25,9 +18,15 @@ module.exports = {
         const nochannelEmbed = new Discord.MessageEmbed()
             .setColor(`RED`)
             .setTitle(`⛔ Broadcast channel not found!`)
-
+        const toolongtitleEmbed = new Discord.MessageEmbed()
+            .setColor(`RED`)
+            .setTitle(`⛔ Title can't be longer than 256 chatacters`)
+        const toolongdescriptionEmbed = new Discord.MessageEmbed()
+            .setColor(`RED`)
+            .setTitle(`⛔ Description can't be longer than 2048 chatacters`)
+        
         if (!msg.member.hasPermission("ADMINISTRATOR")) return msg.reply(nopermEmbed).then(msg => msg.delete({ timeout: 5000 }));
-
+        
         const titleembed = new Discord.MessageEmbed()
             .setColor('GREEN')
             .setTitle("🔴Broadcast")
@@ -43,19 +42,23 @@ module.exports = {
             .setTitle("🔴 Broadcast")
             .setDescription("OUT OF TIME!")
 
-        msg.channel.send(titleembed)
+        const filter = m => m.author.id === msg.author.id;
+
+        msg.channel.send(titleembed)           
             
             .then(msg => {//collect title
             msg.channel.awaitMessages(filter, {max: 1,time: 60000,errors: ['time']})
                 .then(collected => {
-                title = collected.first().content
+                var title = collected.first().content;
+                if (title.length > 256) return msg.channel.send(toolongtitleEmbed).then(msg => msg.delete({ timeout: 5000 }));
                 collected.first().delete();
-                msg.edit(descriptionembed)
+                msg.edit(descriptionembed)                  
                     
                     .then(msg => {//collect description
                     msg.channel.awaitMessages(filter, {max: 1,time: 60000,errors: ['time']})
                         .then(collected => {
-                        description = collected.first().content;
+                        var description = collected.first().content;
+                        if (description.length > 256) return msg.channel.send(toolongdescriptionEmbed).then(msg => msg.delete({ timeout: 5000 }));
                         collected.first().delete();
                         msg.delete();
                             
@@ -64,14 +67,12 @@ module.exports = {
                                     .setTitle(title)
                                     .setDescription(description)
                                     .setColor("#00ff00")
-                                    .setAuthor(author.username, author.displayAvatarURL())    
+                                    .setAuthor(msg.author.username, msg.author.displayAvatarURL())    
                                 let bcchannel = msg.guild.channels.cache.find(bcchannel => bcchannel.id === (client.settings.get(msg.guild.id, "bcchannel")));
                                 if(!bcchannel) return msg.channel.send(nochannelEmbed).then(msg => msg.delete({ timeout: 5000 }));
                                 bcchannel.send("@everyone");
                                 bcchannel.send(bcEmbed);
-                                } 
-
-                                catch(err) {//bot no perm error
+                                } catch(err) {//bot no perm error
                                 msg.channel.send(nobotpermEmbed).then(msg => msg.delete({ timeout: 5000 }));
                                 console.log(err);
                                 }
