+++ commands/Admin/giveaway.js
@@ -9,12 +9,7 @@ module.exports = {
     usage: "giveaway\n**e.g.**\n\`giveaway\`\n> Follow the instructions\n> Bot will ask you to enter a time and a prize",
     permission: "MANAGE_GUILD",
     run: async (client, msg, arg) => {
-        if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
     
-        var duration;
-        var prize;
-        var filter = m => m.author.id === msg.author.id;
-
         const nopermEmbed = new Discord.MessageEmbed()
             .setColor(`RED`)
             .setTitle(`⛔ You don't have permission to use this!`)
@@ -39,18 +34,24 @@ module.exports = {
         const outoftimeEmbed = new Discord.MessageEmbed()
             .setColor(`RED`)
             .setTitle(`⛔ No answer after 30 seconds, operation canceled.`)
-
+        const toolongEmbed = new Discord.MessageEmbed()
+            .setColor(`RED`)
+            .setTitle(`⛔ Prize can't be longer than 240 chatacters`)
+        
         if(!msg.guild.member(msg.author).hasPermission('MANAGE_GUILD')) return msg.channel.send(nopermEmbed).then(msg => msg.delete({ timeout: 5000 }));
         let gachannel = msg.guild.channels.cache.find(gachannel => gachannel.id === (client.settings.get(msg.guild.id, "gachannel")));
         if(!gachannel) return msg.reply(nochannelEmbed).then(msg => msg.delete({ timeout: 5000 }));
+
+        var filter = m => m.author.id === msg.author.id;
         msg.channel.send(sendtimeEmbed).then(msg => {//collect duration
             msg.channel.awaitMessages(filter, {max: 1,time: 30000,errors: ['time']}).then(collected => {
                 if(!collected.first().content.match(/[1-60][s,m,h,d,w]/g)) return msg.channel.send(invalidtimeEmbed).then(msg => msg.delete({ timeout: 5000 }));
-                duration = collected.first().content
+                var duration = collected.first().content
                 collected.first().delete();
                 msg.edit(sendprizeEmbed).then(msg => {//collect prize
                     msg.channel.awaitMessages(filter, {max: 1,time: 30000,errors: ['time']}).then(collected => {
-                        prize = collected.first().content;
+                        var prize = collected.first().content;
+                        if (prize.length > 240) return msg.channel.send(toolongEmbed).then(msg => msg.delete({ timeout: 5000 }));
                         collected.first().delete();
                         msg.delete();
                         try {
@@ -62,7 +63,7 @@ module.exports = {
                             .setFooter("Created at:")
                             .setTimestamp()
                         gachannel.send(startEmbed).then(m => {
-                            let re = m.react('🎉');
+                            m.react('🎉');
                             //update time left
                             let live = ms(duration);
                             var livetest = setInterval (liveupdate, 30000);
@@ -99,9 +100,9 @@ module.exports = {
                                 gachannel.send(`**Congratulations ${winner}!\nYou won: \`${prize}\`**`)
                             }, ms(duration));
                         });
-                        } catch(e) {
+                        } catch(err) {
                             msg.channel.send(nobotpermEmbed);
-                            console.log(e);
+                            console.log(err);
                         }
                     }).catch(() => {return msg.edit(outoftimeEmbed).then(msg => msg.delete({ timeout: 5000 }));});
                 });
